ＷＥＢアプリ開発２　１－５．BookApp.vueからコンテンツ分割初版
